Add new option to be able to ignore inner dimensions check
When the library checks whether an element is navigable or not it checks
whether the inner width and height values are greater than zero. There're
cases where the inner width and height are zero but the element is still visible
and therefore needs to be focusable. To fix it I have added a new option
to disable that check for a particular section or globally.